+++ package.json
@@ -13,8 +13,8 @@
   },
   "dependencies": {
     "axios": "^0.21.0",
+    "cheerio-without-node-native": "^1.0.0-rc.2",
     "core-js": "^3.6.5",
-    "genius-lyrics-api": "^3.0.3",
     "vue": "^2.6.11",
     "vue-router": "^3.4.9",
     "vuetify": "^2.2.11",
